Set up SWIG bindings over the whole library
